Revert "engine: sync only own device before snapshot"
This reverts commit da2a02e86462fb4f3abac4d0f91f2380ff34053c.

The commit (da2a02e) results in a regression that v2 volume data are not sync before taking a snapshot.
Using sync directly with a block device path does not work as expected.

According to the doc of the sync command, what it does is to synchronize the filesystem which contains the specified file.

Longhorn 8977

Signed-off-by: Derek Su <derek.su@suse.com>