fix: reconstruct the cache for failed replica restarting
If the upper layer tries to launch a previously failed replica
in a newly launched instance manager pod. The pod may have no
info then directly treat it as a new replica and skip the cache
reconstruction. Finally, this replica will be failed by
r.validateAndUpdate() in the next moment.

Longhorn 9488

Signed-off-by: Shuo Wu <shuo.wu@suse.com>